добавил функции
