+++ htmlcov/index.html
@@ -11,7 +11,7 @@
 <header>
     <div class="content">
         <h1>Coverage report:
-            <span class="pc_cov">90%</span>
+            <span class="pc_cov">18%</span>
         </h1>
         <aside id="help_panel_wrapper">
             <input id="help_panel_state" type="checkbox">
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.8.0">coverage.py v7.8.0</a>,
-            created at 2025-04-21 04:08 +0700
+            created at 2025-04-21 20:10 +0700
         </p>
     </div>
 </header>
@@ -80,16 +80,16 @@
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_api_utils_py.html">src\api_utils.py</a></td>
                 <td>63</td>
+                <td>50</td>
                 <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="63 63">100%</td>
+                <td class="right" data-ratio="13 63">21%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_config_py.html">src\config.py</a></td>
-                <td>15</td>
+                <td>17</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="15 15">100%</td>
+                <td class="right" data-ratio="17 17">100%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_loggers_py.html">src\loggers.py</a></td>
@@ -100,40 +100,61 @@
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_reports_py.html">src\reports.py</a></td>
-                <td>8</td>
-                <td>8</td>
+                <td>48</td>
+                <td>48</td>
                 <td>0</td>
-                <td class="right" data-ratio="0 8">0%</td>
+                <td class="right" data-ratio="0 48">0%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_reports_utils_py.html">src\reports_utils.py</a></td>
+                <td>21</td>
+                <td>21</td>
+                <td>0</td>
+                <td class="right" data-ratio="0 21">0%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_services_py.html">src\services.py</a></td>
-                <td>11</td>
-                <td>11</td>
+                <td>17</td>
+                <td>17</td>
                 <td>0</td>
-                <td class="right" data-ratio="0 11">0%</td>
+                <td class="right" data-ratio="0 17">0%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_services_utils_py.html">src\services_utils.py</a></td>
+                <td>24</td>
+                <td>24</td>
+                <td>0</td>
+                <td class="right" data-ratio="0 24">0%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_transaction_utils_py.html">src\transaction_utils.py</a></td>
+                <td>30</td>
+                <td>30</td>
+                <td>0</td>
+                <td class="right" data-ratio="0 30">0%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_utils_py.html">src\utils.py</a></td>
-                <td>135</td>
+                <td>101</td>
+                <td>83</td>
                 <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="135 135">100%</td>
+                <td class="right" data-ratio="18 101">18%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_views_py.html">src\views.py</a></td>
-                <td>4</td>
-                <td>4</td>
+                <td>10</td>
+                <td>10</td>
                 <td>0</td>
-                <td class="right" data-ratio="0 4">0%</td>
+                <td class="right" data-ratio="0 10">0%</td>
             </tr>
         </tbody>
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>252</td>
-                <td>24</td>
+                <td>347</td>
+                <td>284</td>
                 <td>0</td>
-                <td class="right" data-ratio="228 252">90%</td>
+                <td class="right" data-ratio="63 347">18%</td>
             </tr>
         </tfoot>
     </table>
@@ -145,7 +166,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.8.0">coverage.py v7.8.0</a>,
-            created at 2025-04-21 04:08 +0700
+            created at 2025-04-21 20:10 +0700
         </p>
     </div>
     <aside class="hidden">
