+++ htmlcov/status.json
@@ -1,1 +1,1 @@
-{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.8.0","globals":"c568be692e52ceec26b7afc4c0f5ab28","files":{"z_145eef247bfb46b6___init___py":{"hash":"e6baa73cda2916dad605215f937a92e1","index":{"url":"z_145eef247bfb46b6___init___py.html","file":"src\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":0,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_config_py":{"hash":"76d920d1a831fa280b3f9fd736f19561","index":{"url":"z_145eef247bfb46b6_config_py.html","file":"src\\config.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":15,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_loggers_py":{"hash":"41ec39bf15b32851d13e6667a1fa138d","index":{"url":"z_145eef247bfb46b6_loggers_py.html","file":"src\\loggers.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":16,"n_excluded":0,"n_missing":1,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_reports_py":{"hash":"bd6afda9a556c4ffca15bc8c4683d0e2","index":{"url":"z_145eef247bfb46b6_reports_py.html","file":"src\\reports.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":8,"n_excluded":0,"n_missing":8,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_services_py":{"hash":"898214ee0e278ab7d63046193752bf2b","index":{"url":"z_145eef247bfb46b6_services_py.html","file":"src\\services.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":11,"n_excluded":0,"n_missing":11,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_utils_py":{"hash":"3e96a69dabedda3c2bf390bc73ff6d36","index":{"url":"z_145eef247bfb46b6_utils_py.html","file":"src\\utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":135,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_views_py":{"hash":"6597d5e524043741e073643eb4748632","index":{"url":"z_145eef247bfb46b6_views_py.html","file":"src\\views.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":4,"n_excluded":0,"n_missing":4,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_api_utils_py":{"hash":"606cd0839973148bbedbd54ce366d17a","index":{"url":"z_145eef247bfb46b6_api_utils_py.html","file":"src\\api_utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":63,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
+{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.8.0","globals":"c568be692e52ceec26b7afc4c0f5ab28","files":{"z_145eef247bfb46b6___init___py":{"hash":"e6baa73cda2916dad605215f937a92e1","index":{"url":"z_145eef247bfb46b6___init___py.html","file":"src\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":0,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_config_py":{"hash":"dbf42fd008222ebd3946831e1e555bfd","index":{"url":"z_145eef247bfb46b6_config_py.html","file":"src\\config.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":17,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_loggers_py":{"hash":"41ec39bf15b32851d13e6667a1fa138d","index":{"url":"z_145eef247bfb46b6_loggers_py.html","file":"src\\loggers.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":16,"n_excluded":0,"n_missing":1,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_reports_py":{"hash":"b634e24fc2a7b6451a006e2d61aaa886","index":{"url":"z_145eef247bfb46b6_reports_py.html","file":"src\\reports.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":48,"n_excluded":0,"n_missing":48,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_services_py":{"hash":"3b1ee4a4928e1be72bdc849d40fe5434","index":{"url":"z_145eef247bfb46b6_services_py.html","file":"src\\services.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":17,"n_excluded":0,"n_missing":17,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_utils_py":{"hash":"692f5e347e8c5c639ff6f901396cb4a3","index":{"url":"z_145eef247bfb46b6_utils_py.html","file":"src\\utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":101,"n_excluded":0,"n_missing":83,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_views_py":{"hash":"ca70f8d8d1103f708f0cfee600b0c9c9","index":{"url":"z_145eef247bfb46b6_views_py.html","file":"src\\views.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":10,"n_excluded":0,"n_missing":10,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_api_utils_py":{"hash":"3749560441f032986c8f7c9f4a35d548","index":{"url":"z_145eef247bfb46b6_api_utils_py.html","file":"src\\api_utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":63,"n_excluded":0,"n_missing":50,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_reports_utils_py":{"hash":"fb9582896645a8d1853a20105e74f7e3","index":{"url":"z_145eef247bfb46b6_reports_utils_py.html","file":"src\\reports_utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":21,"n_excluded":0,"n_missing":21,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_services_utils_py":{"hash":"300ceb595c3fc4b7caff5ff4bb18b26d","index":{"url":"z_145eef247bfb46b6_services_utils_py.html","file":"src\\services_utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":24,"n_excluded":0,"n_missing":24,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_transaction_utils_py":{"hash":"9fa0b5888f8c3019e211712747e18436","index":{"url":"z_145eef247bfb46b6_transaction_utils_py.html","file":"src\\transaction_utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":30,"n_excluded":0,"n_missing":30,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
