+++ htmlcov/z_145eef247bfb46b6_config_py.html
@@ -54,8 +54,8 @@
             </div>
         </aside>
         <h2>
-            <span class="text">15 statements &nbsp;</span>
-            <button type="button" class="run button_toggle_run" value="run" data-shortcut="r" title="Toggle lines run">15<span class="text"> run</span></button>
+            <span class="text">17 statements &nbsp;</span>
+            <button type="button" class="run button_toggle_run" value="run" data-shortcut="r" title="Toggle lines run">17<span class="text"> run</span></button>
             <button type="button" class="mis show_mis button_toggle_mis" value="mis" data-shortcut="m" title="Toggle lines missing">0<span class="text"> missing</span></button>
             <button type="button" class="exc show_exc button_toggle_exc" value="exc" data-shortcut="x" title="Toggle lines excluded">0<span class="text"> excluded</span></button>
         </h2>
@@ -65,7 +65,7 @@
             <a id="nextFileLink" class="nav" href="z_145eef247bfb46b6_loggers_py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
             <a class="nav" href="https://coverage.readthedocs.io/en/7.8.0">coverage.py v7.8.0</a>,
-            created at 2025-04-21 03:30 +0700
+            created at 2025-04-21 20:10 +0700
         </p>
         <aside class="hidden">
             <button type="button" class="button_next_chunk" data-shortcut="j"></button>
@@ -96,6 +96,8 @@
     <p class="run"><span class="n"><a id="t14" href="#t14">14</a></span><span class="t"><span class="nam">FILE_USER_SETTINGS</span> <span class="op">=</span> <span class="str">"user_settings.json"</span>&nbsp;</span><span class="r"></span></p>
     <p class="run"><span class="n"><a id="t15" href="#t15">15</a></span><span class="t"><span class="nam">FILE_OPERATIONS</span> <span class="op">=</span> <span class="str">"operations.xlsx"</span>&nbsp;</span><span class="r"></span></p>
     <p class="run"><span class="n"><a id="t16" href="#t16">16</a></span><span class="t"><span class="nam">DATA_FOLDER_NAME</span> <span class="op">=</span> <span class="str">"data"</span>&nbsp;</span><span class="r"></span></p>
+    <p class="run"><span class="n"><a id="t17" href="#t17">17</a></span><span class="t"><span class="nam">REPORTS_FOLDER_NAME</span> <span class="op">=</span> <span class="str">"reports"</span>&nbsp;</span><span class="r"></span></p>
+    <p class="run"><span class="n"><a id="t18" href="#t18">18</a></span><span class="t"><span class="nam">RUSSIAN_DAYS</span> <span class="op">=</span> <span class="op">[</span><span class="str">"&#1055;&#1086;&#1085;&#1077;&#1076;&#1077;&#1083;&#1100;&#1085;&#1080;&#1082;"</span><span class="op">,</span> <span class="str">"&#1042;&#1090;&#1086;&#1088;&#1085;&#1080;&#1082;"</span><span class="op">,</span> <span class="str">"&#1057;&#1088;&#1077;&#1076;&#1072;"</span><span class="op">,</span> <span class="str">"&#1063;&#1077;&#1090;&#1074;&#1077;&#1088;&#1075;"</span><span class="op">,</span> <span class="str">"&#1055;&#1103;&#1090;&#1085;&#1080;&#1094;&#1072;"</span><span class="op">,</span> <span class="str">"&#1057;&#1091;&#1073;&#1073;&#1086;&#1090;&#1072;"</span><span class="op">,</span> <span class="str">"&#1042;&#1086;&#1089;&#1082;&#1088;&#1077;&#1089;&#1077;&#1085;&#1100;&#1077;"</span><span class="op">]</span>&nbsp;</span><span class="r"></span></p>
 </main>
 <footer>
     <div class="content">
@@ -105,7 +107,7 @@
             <a class="nav" href="z_145eef247bfb46b6_loggers_py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
             <a class="nav" href="https://coverage.readthedocs.io/en/7.8.0">coverage.py v7.8.0</a>,
-            created at 2025-04-21 03:30 +0700
+            created at 2025-04-21 20:10 +0700
         </p>
     </div>
 </footer>
