+++ tests/test_utils.py
@@ -4,10 +4,8 @@ from unittest.mock import mock_open, patch
 
 import pytest
 
-from src.utils import (get_amount_for_categories, get_cashback_categories, get_greetings, get_invest_amount,
-                       get_json_file, get_last_digits_card_number, get_search_by_keyword, get_start_date,
-                       get_total_amount, get_total_amount_for_card, get_transactions_by_date_period,
-                       get_transactions_for_categories, top_transactions_by_amount)
+from src.utils import (get_amount_for_categories, get_greetings, get_json_file, get_last_digits_card_number,
+                       get_start_date, get_total_amount, get_total_amount_for_card)
 
 
 @pytest.mark.parametrize(
@@ -47,106 +45,6 @@ def test_get_start_date(date, period, expected):
     assert get_start_date(date, period) == expected
 
 
-def test_get_transactions_by_date_period(list_transactions, tr_by_period):
-    assert (
-        get_transactions_by_date_period(list_transactions, datetime(2018, 1, 3), datetime(2018, 1, 5)) == tr_by_period
-    )
-
-
-def test_get_transactions_by_date_period_rev(tr_by_period, test_date_start, test_date_end):
-    assert get_transactions_by_date_period([{}], test_date_end, test_date_start) == []
-
-
-def test_top_transactions_by_amount(list_transactions):
-    assert top_transactions_by_amount(list_transactions) == [
-        {
-            "Дата операции": "01.01.2018 12:49:53",
-            "Дата платежа": "01.01.2018",
-            "Номер карты": None,
-            "Статус": "OK",
-            "Сумма операции": -3000.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -3000.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Переводы",
-            "MCC": None,
-            "Описание": "Линзомат ТЦ Юность",
-            "Бонусы (включая кэшбэк)": 0,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 3000.0,
-        },
-        {
-            "Дата операции": "01.01.2018 20:27:51",
-            "Дата платежа": "04.01.2018",
-            "Номер карты": "*7197",
-            "Статус": "OK",
-            "Сумма операции": -316.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -316.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Красота",
-            "MCC": 5977.0,
-            "Описание": "OOO Balid",
-            "Бонусы (включая кэшбэк)": 6,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 316.0,
-        },
-        {
-            "Дата операции": "03.01.2018 15:03:35",
-            "Дата платежа": "04.01.2018",
-            "Номер карты": "*7197",
-            "Статус": "OK",
-            "Сумма операции": -73.06,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -73.06,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Супермаркеты",
-            "MCC": 5499.0,
-            "Описание": "Magazin 25",
-            "Бонусы (включая кэшбэк)": 1,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 73.06,
-        },
-        {
-            "Дата операции": "03.01.2018 14:55:21",
-            "Дата платежа": "05.01.2018",
-            "Номер карты": "*7197",
-            "Статус": "OK",
-            "Сумма операции": -21.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -21.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Красота",
-            "MCC": 5977.0,
-            "Описание": "OOO Balid",
-            "Бонусы (включая кэшбэк)": 0,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 21.0,
-        },
-        {
-            "Дата операции": "23.01.2018 14:55:21",
-            "Дата платежа": "25.01.2018",
-            "Номер карты": "*7197",
-            "Статус": "OK",
-            "Сумма операции": 21.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": 21.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Красота",
-            "MCC": 5977.0,
-            "Описание": "OOO Balid",
-            "Бонусы (включая кэшбэк)": 0,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 21.0,
-        },
-    ]
-
-
 @patch("src.utils.open")
 def test_get_json_file_err_open(mock_file):
     mock_file.side_effect = FileNotFoundError
@@ -183,110 +81,3 @@ def test_get_amount_for_categories(list_transactions):
 
 def test_get_amount_for_categories_without_cats(list_transactions_without_cats):
     assert get_amount_for_categories(list_transactions_without_cats) == {"Переводы": 3000.0, "Супермаркеты": 73.06}
-
-
-def test_get_transactions_for_categories(list_transactions):
-    assert get_transactions_for_categories(list_transactions, {"Красота", "Переводы"}) == [
-        {
-            "Дата операции": "01.01.2018 20:27:51",
-            "Дата платежа": "04.01.2018",
-            "Номер карты": "*7197",
-            "Статус": "OK",
-            "Сумма операции": -316.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -316.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Красота",
-            "MCC": 5977.0,
-            "Описание": "OOO Balid",
-            "Бонусы (включая кэшбэк)": 6,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 316.0,
-        },
-        {
-            "Дата операции": "01.01.2018 12:49:53",
-            "Дата платежа": "01.01.2018",
-            "Номер карты": None,
-            "Статус": "OK",
-            "Сумма операции": -3000.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -3000.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Переводы",
-            "MCC": None,
-            "Описание": "Линзомат ТЦ Юность",
-            "Бонусы (включая кэшбэк)": 0,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 3000.0,
-        },
-        {
-            "Дата операции": "03.01.2018 14:55:21",
-            "Дата платежа": "05.01.2018",
-            "Номер карты": "*7197",
-            "Статус": "OK",
-            "Сумма операции": -21.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -21.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Красота",
-            "MCC": 5977.0,
-            "Описание": "OOO Balid",
-            "Бонусы (включая кэшбэк)": 0,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 21.0,
-        },
-        {
-            "Дата операции": "23.01.2018 14:55:21",
-            "Дата платежа": "25.01.2018",
-            "Номер карты": "*7197",
-            "Статус": "OK",
-            "Сумма операции": 21.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": 21.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Красота",
-            "MCC": 5977.0,
-            "Описание": "OOO Balid",
-            "Бонусы (включая кэшбэк)": 0,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 21.0,
-        },
-    ]
-
-
-def test_get_cashback_categories(list_transactions):
-    assert get_cashback_categories(list_transactions, percent_cashback=5.0) == {
-        "Переводы": 150,
-        "Красота": 16,
-        "Супермаркеты": 3,
-    }
-
-
-def test_get_invest_amount(tr_by_period):
-    assert get_invest_amount(tr_by_period, 50) == 55.94
-
-
-def test_get_search_by_keyword(list_transactions):
-    assert get_search_by_keyword(list_transactions, "перевод", {"Категория"}) == [
-        {
-            "Дата операции": "01.01.2018 12:49:53",
-            "Дата платежа": "01.01.2018",
-            "Номер карты": None,
-            "Статус": "OK",
-            "Сумма операции": -3000.0,
-            "Валюта операции": "RUB",
-            "Сумма платежа": -3000.0,
-            "Валюта платежа": "RUB",
-            "Кэшбэк": None,
-            "Категория": "Переводы",
-            "MCC": None,
-            "Описание": "Линзомат ТЦ Юность",
-            "Бонусы (включая кэшбэк)": 0,
-            "Округление на инвесткопилку": 0,
-            "Сумма операции с округлением": 3000.0,
-        }
-    ]
